+++ canopy/definitions/base.py
@@ -11,9 +11,9 @@ class KindField:
     """
     klass: type(forms.Field)
     # Previously named 'options'
-    attributes_initials: dict = field(default_factory=dict)
+    initials: dict = field(default_factory=dict)
     # Previously named 'options_fields'
-    attributes_fields: dict = field(default_factory=dict)
+    options: dict = field(default_factory=dict)
 
 
 @dataclass(frozen=True)
@@ -22,11 +22,11 @@ class KindWidget:
     Kind field declare what Django form widget to use for a Django form a field from
     a KindField.
 
-    Opposed to KindField, there is no ``attributes_initials`` since a Widget itself does
+    Opposed to KindField, there is no ``initials`` since a Widget itself does
     not have initial value.
     """
     klass: type(forms.MediaDefiningClass)
-    attributes_fields: dict = field(default_factory=dict)
+    options: dict = field(default_factory=dict)
 
 
 @dataclass(frozen=True)