+++ canopy/definitions/kinds/__init__.py
@@ -1,0 +1,10 @@
+from .boolean import BooleanKind
+from .text import EmailKind, TextKind, TextareaKind
+
+
+__all__ = [
+    "BooleanKind",
+    "EmailKind",
+    "TextKind",
+    "TextareaKind",
+]