+++ canopy/definitions/kinds/text.py
@@ -1,14 +1,7 @@
 from django import forms
 from django.utils.translation import gettext_lazy as _
 
-from .slots import Kind, KindField, KindWidget
-
-
-BooleanKind = Kind(
-    identifier="boolean",
-    name="Boolean",
-    field=KindField(klass=forms.BooleanField)
-)
+from ..base import Kind, KindField, KindWidget
 
 
 TextKind = Kind(
@@ -16,8 +9,8 @@ TextKind = Kind(
     name="Simple text",
     field=KindField(
         klass=forms.CharField,
-        attributes_initials={"max_length": 255},
-        attributes_fields={
+        initials={"max_length": 255},
+        options={
             "max_length": forms.IntegerField(
                 min_value=1,
                 max_value=255,
@@ -34,7 +27,7 @@ EmailKind = Kind(
     name="Email",
     field=KindField(
         klass=forms.EmailField,
-        attributes_fields={
+        options={
             "max_length": forms.IntegerField(
                 min_value=1,
                 max_value=255,
@@ -50,7 +43,7 @@ TextareaKind = Kind(
     name="Textarea",
     field=KindField(
         klass=forms.CharField,
-        attributes_initials={"max_length": 3000}
+        initials={"max_length": 3000}
     ),
     widget=KindWidget(klass=forms.Textarea)
 )