+++ CHANGELOG.rst
@@ -3,12 +3,19 @@
 Changelog
 =========
 
+Version 0.3.0 - Unreleased
+--------------------------
+
+* Refactored Slot definition with dataclasses;
+
+
 Version 0.2.0 - Unreleased
 --------------------------
 
 * The controller form save is now working with basic fields;
 * Added working basic controller form;
 * Added model admins;
+* Refactored registry;
 
 
 Version 0.1.0 - Unreleased